requirements.txt
streamlit
tensorflow==2.17.0
numpy
pillow
joblib
scikit-learn
